+++ R/smooth_percentile.R
@@ -7,20 +7,20 @@
 #'
 #' @param data The data given to this function during the calculations
 #' performed by \code{\link{ts2clm}}.
-#' @param smooth_percentile_width The width of the smoothing window
+#' @param smoothPercentileWidth The width of the smoothing window
 #' to be applied.
 #'
 #' @return The function returns the data in the same format it was
 #' input as, with the climatology values smoothed as desired.
-smooth_percentile <- function(data, smooth_percentile_width) {
+smooth_percentile <- function(data, smoothPercentileWidth) {
 
-  seas_clim_year <- thresh_clim_year <- var_clim_year <- NULL
+  seas <- thresh <- var <- NULL
 
   clim <- data %>%
     dplyr::mutate(
-      seas_clim_year = raster::movingFun(
-        seas_clim_year,
-        n = smooth_percentile_width,
+      seas = raster::movingFun(
+        seas,
+        n = smoothPercentileWidth,
         fun = mean,
         type = "around",
         circular = TRUE,
@@ -28,9 +28,9 @@ smooth_percentile <- function(data, smooth_percentile_width) {
       )
     ) %>%
     dplyr::mutate(
-      thresh_clim_year = raster::movingFun(
-        thresh_clim_year,
-        n = smooth_percentile_width,
+      thresh = raster::movingFun(
+        thresh,
+        n = smoothPercentileWidth,
         fun = mean,
         type = "around",
         circular = TRUE,
@@ -38,9 +38,9 @@ smooth_percentile <- function(data, smooth_percentile_width) {
       )
     ) %>%
     dplyr::mutate(
-      var_clim_year = raster::movingFun(
-        var_clim_year,
-        n = smooth_percentile_width,
+      var = raster::movingFun(
+        var,
+        n = smoothPercentileWidth,
         fun = mean,
         type = "around",
         circular = TRUE,